Improve dynamicset informers concurrency
Currently, when starting and stopping informers, the informer map is
updated in a non-safe manner.

This means there might be a race condition leaking informers that would
become unstoppables.

In particular, this sequence may happen

```
// routine 1             routine 2
Register() {

informers.Load() // not loaded
                         Register(){

                         informers.Load() // not loaded
Start() // this informer is leaked and unstoppable

informers.Store()
                         Start()
                         informers.Store() // this informer replaces the previous one

}
                         }
```

With this commit, we use native proven map functions to avoid leaking
informers

We expect this to happen

// routine 1             routine 2
Register() {

informers.LoadOrStore() // not loaded
                         Register(){
Start()
                         informers.LoadOrStore() // loaded
}
                         }