Decouple informer management and handler registration
As described in comments of #323, there is a need to trigger
RGD reconciliation when managed resources are updated.

To do so, we envision decoupling dynamic handlers and informers.

That way we will be able to start informers for all the
managed resources and map changes to the managed objects
to trigger a RGD reconciliation