Create ready-to-install kustomizations
As described in #729, there is a desire to install kro
with a plain kubectl apply (or kustomize)

This commit is the first step to achieve this goal.

It enables installing kro with and without prometheus metrics
directly with kubectl.

The simplest way to achieve the whole experience is to
upload the generated manifests to the github release.

Alternatively, there can be an automation to update the manifests
as part of the release process.

Manifests are generated with

```
kubectl kustomize  ./manifests/core-install/
kubectl kustomize  ./manifests/core-install-with-metrics
```