Create ready-to-install kustomizations
As described in #729, there is a desire to install kro
with a plain kubectl apply (or kustomize)

This commit is the first step to achieve this goal.

It enables installing kro without prometheus metrics
directly with kubectl. and removes unnecessary manifests created by kube-builder.

The simplest way to achieve the whole experience is to,
later, upload the generated manifests to the github release.

This installation mimics `helm install --set rbac.mode=aggregation`
It makes it a significant change compared to HELM where the default mode
is `unrestricted`.

Manifests are generated with

```
kubectl kustomize  ./manifests/core-install/
```