Configure kube-builder to generate helm-compatible RBAC role
As part of #729, we want to create the same deployment in plain
manifests as what we have in helm.

The current implementation has changed since the bootstrap of the project

In helm, we have both "aggregated" and "unrestricted" RBAC modes

This commit aims at enabling backporting thos 2 modes in kustomize.